backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
passlib
itsdangerous
